server/requirements.txt
Django==4.1.1
psycopg2==2.9.3
celery==5.2.7
redis==4.3.4
djangorestframework==3.13.1
drf-spectacular==0.23.1
biopython==1.79

